requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bitsandbytes==0.39.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cpm-kernels==1.0.11
cycler==0.11.0
datasets==2.11.0
deepspeed==0.14.0
dill==0.3.6
einops==0.7.0
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.10.7
Flask==2.3.2
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2024.3.1
future==0.18.3
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gradio==3.24.1
gradio_client==0.0.7
grpcio==1.59.0
h11==0.14.0
hjson==3.1.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.21.4
icetk==0.0.4
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
lit==16.0.0
Markdown==3.5
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nebula2-python==2.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
orjson==3.8.9
pandas==2.0.0
peft==0.5.0
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.0
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==19.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rouge-chinese==1.0.3
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.32.0
transformers-stream-generator==0.0.4
triton==2.0.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.1
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
